Revert "feat: improve login feature"
This reverts commit ab2d4e62db7060a492a66162469281481792a34d.
